refactor auth exception
