+++ src/main/java/com/khutircraftubackend/auth/exception/AuthenticationException.java
@@ -1,12 +1,12 @@
-package com.khutircraftubackend.auth;
+package com.khutircraftubackend.auth.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.UNAUTHORIZED)
-public class CustomAuthenticationException extends RuntimeException {
+public class AuthenticationException extends RuntimeException {
 
-    public CustomAuthenticationException(String message){
+    public AuthenticationException(String message){
         super (message);
     }
 }
